+++ packages/contracts/src/gateway/post/response/index.ts
@@ -1,1 +1,2 @@
 export * from './post.response.js';
+export * from './comment.response.js';